pyproject.toml
[project]
name = "agent-openai-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=25.1.0",
    "openai-agents>=0.0.16",
    "python-dotenv>=1.1.0",
    "resend>=2.10.0",
]
